test flutter v6.0.0 applications

+++ flutter/test-app-pub/lib/pipe_tab.dart
@@ -160,7 +160,7 @@ class PipeTab implements PlayerTab {
       return;
     }
 
-    int timeInMilliseconds = this._statistics!.getTime();
+    double timeInMilliseconds = this._statistics!.getTime();
     int totalVideoDuration = 9000;
 
     int completePercentage = (timeInMilliseconds * 100) ~/ totalVideoDuration;
